# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Book Management System API
  description: |
    ## Comprehensive Book Management API
    
    This is a fully-featured REST API for managing books and user favorites with advanced search capabilities.
    
    ### Features
    - **Complete CRUD Operations** for books and favorites
    - **Advanced Search** with relevance scoring and multiple search strategies
    - **Search Suggestions** for autocomplete functionality
    - **Input Validation** with detailed error messages
    - **Consistent API Response Format** across all endpoints
    - **Soft Delete** support for books
    - **Duplicate Prevention** for books and favorites
    
    ### Search Capabilities
    - **Exact Match**: Find exact title or author matches
    - **Prefix Search**: Find titles/authors starting with query
    - **Contains Search**: Find titles/authors containing query (default)
    - **Fuzzy Search**: Advanced matching with typo tolerance
    - **Relevance Scoring**: Results ranked by relevance
    - **Multi-field Search**: Search across title, author, and category
    
    ### API Response Format
    All endpoints return a consistent response structure:
    ```json
    {
      "success": true,
      "message": "Operation completed successfully",
      "data": { ... }
    }
    ```
    
    Error responses:
    ```json
    {
      "success": false,
      "message": "Error description",
      "error": "Detailed error message"
    }
    ```
  version: 2.0.0
  contact:
...
servers:
...
paths:
...
  /favorites:
    get:
      tags:
        - Favorites
      summary: Get all user favorites
      description: |
        Retrieve a list of all books marked as favorites by the current user. 
        Each favorite includes book details such as title, author, and category. 
        The user ID is currently hardcoded to 1 for demo purposes.
      responses:
        '200':
          description: Favorites retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteListResponse'
              example:
                success: true
                message: "Favorites retrieved successfully"
                data:
                  - id: 1
                    user_id: 1
                    book_id: 1
                    created_at: "2023-01-01T00:00:00Z"
                    book:
                      id: 1
                      title: "Harry Potter and the Sorcerer's Stone"
                      author: "J.K. Rowling"
                      category: "Fantasy"
                      created_at: "2023-01-01T00:00:00Z"
                      updated_at: "2023-01-01T00:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Favorites
      summary: Add book to favorites
      description: |
        Add a book to the current user's favorites list. Validates that the book exists and checks for duplicates. 
        The user ID is currently hardcoded to 1 for demo purposes. Returns the complete favorite details including book information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteRequest'
            example:
              book_id: 1
      responses:
        '201':
          description: Favorite added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
              example:
                success: true
                message: "Favorite added successfully"
                data:
                  id: 1
                  user_id: 1
                  book_id: 1
                  created_at: "2023-01-01T00:00:00Z"
                  book:
                    id: 1
                    title: "Harry Potter and the Sorcerer's Stone"
                    author: "J.K. Rowling"
                    category: "Fantasy"
                    created_at: "2023-01-01T00:00:00Z"
                    updated_at: "2023-01-01T00:00:00Z"
        '400':
          description: Invalid request body or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Book already in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/{id}:
    get:
      tags:
        - Favorites
      summary: Get favorite by ID
      description: |
        Retrieve detailed information about a specific favorite entry using its unique identifier. 
        Returns complete favorite details including the associated book information.
      parameters:
        - name: id
          in: path
          required: true
          description: Favorite ID (must be a positive integer)
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Favorite retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
        '400':
          description: Invalid favorite ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Favorites
      summary: Remove book from favorites
      description: |
        Remove a book from the current user's favorites list using the favorite entry ID. 
        This operation only removes the favorite relationship, not the book itself. 
        The user ID is currently hardcoded to 1 for demo purposes.
      parameters:
        - name: id
          in: path
          required: true
          description: Favorite ID (must be a positive integer)
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Favorite removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Favorite removed successfully"
        '400':
          description: Invalid favorite ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Harry Potter and the Sorcerer's Stone"
        author:
          type: string
          example: "J.K. Rowling"
        category:
          type: string
          example: "Fantasy"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...
    Favorite:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        book_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        book:
          $ref: '#/components/schemas/Book'
    FavoriteRequest:
      type: object
      required:
        - book_id
      properties:
        book_id:
          type: integer
          description: "ID of the book to add to favorites"
          minimum: 1
          example: 1
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          description: "Response data (varies by endpoint)"
        error:
          type: string
          description: "Error message (only present when success=false)"
          example: "Detailed error message"
...
    FavoriteResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Favorite'
    FavoriteListResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Favorite'
...
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          example:
            success: true
            message: "Operation completed successfully"
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            message:
              type: string
              example: "Error description"
            error:
              type: string
              example: "Detailed error message"
          example:
            success: false
            message: "Validation failed"
            error: "title: Title is required"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...